Move customer to separate file.
